requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
beautifulsoup4==4.1.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
click==7.1.2
colorama==0.4.3
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
httpie==2.1.0
idna==2.8
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
pipenv==2020.6.2
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
six==1.14.0
soupsieve==1.9.3
urllib3==1.25.3
virtualenv==20.0.25
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zipp==3.1.0
